test: consolidate debug test functions into single file
Merge all debug/helper test functions into debug_helper_test.go
  for better organization and maintainability.

  Changes:
  - Consolidate TestDebugAddress from debug_address_test.go
  - Consolidate TestDebugMemberCheck from test_member_check.go
  - Consolidate TestDebugMembershipSimple from debug_member_test.go
  - All debug test functions now in systemcontracts/test/debug_helper_test.go
